requirements.txt
Flask~=3.1.0
Flask-Migrate~=4.0.7
python-dotenv~=1.0.1
marshmallow~=3.24.1
pytest~=8.3.4
alembic~=1.14.0
SQLAlchemy~=2.0.36